% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate}
\usage{
aggregate(polys, pts, agg = list(c("count", "", "num_of_pts")))
}
\arguments{
\item{polys}{Polygon(s) defining area to aggregate.}

\item{pts}{Points with values to aggregate.}

\item{agg}{Aggregation object. For more information, see details.}
}
\description{
Aggregate
}
\details{
Provides a statistical summary of values from points that are
         within a polygon.   The \code{agg} parameter must be a list of
         vector(s), that contain three values: aggregation, inField, and
         outField.  Aggregation may be a string of one of the following:
         sum, average, count, min, max, or deviation. The inField is the
         name of the field on \code{pts} on which you wish to perfrom the
         aggregation and outField is the name of the field on the ouput
         polygon FeatureCollection that will store the resultant value.
}
\examples{
data(lawn_data)
ex_polys <- lawn_data$polygons_aggregate
ex_pts <- lawn_data$points_aggregate
ex_agg <- list(c("sum","population","pop_sum"),
            c("mean","population","pop_mean"),
            c("count","","num_of_pts"))
aggregate(ex_polys,ex_pts,ex_agg)
}
\author{
Jeff Hollister \email{hollister.jeff@epa.gov}
}

