% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.turf.R
\name{as.turf}
\alias{as.turf}
\title{Coerce to turf.js javascript function call}
\usage{
as.turf(x, ...)
}
\arguments{
\item{x}{Input data to be coerced.}

\item{...}{Ignored.}
}
\description{
Coerce to turf.js javascript function call
}
\keyword{internal}
