% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as_feature.R
\name{as_feature}
\alias{as_feature}
\title{Convert a FeatureCollection to a Feature}
\usage{
as_feature(x)
}
\arguments{
\item{x}{A \code{\link{data-FeatureCollection}}}
}
\description{
Convert a FeatureCollection to a Feature
}
\details{
If there are more than one feature within the featurecollection,
each feature is split out into a separate feature, returned in a list.
Each feature is a assigned a class matching it's GeoJSON data type
(e.g., point, polygon, linestring).
}
\examples{
as_feature(lawn_random())
# as_feature(lawn_random("polygons"))
}

