% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bearing.R
\name{bearing}
\alias{bearing}
\title{Bearing}
\usage{
bearing(start, end, lint = FALSE)
}
\arguments{
\item{start}{starting \code{\link{data-Point}}}

\item{end}{ending \code{\link{data-Point}}}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
bearing (numeric) in decimal degrees
}
\description{
Bearing
}
\details{
Takes two points and finds the geographic bearing between them
}
\examples{
start <- '{
 "type": "Feature",
 "properties": {
   "marker-color": "#f00"
 },
 "geometry": {
   "type": "Point",
   "coordinates": [-75.343, 39.984]
 }
}'

end <- '{
  "type": "Feature",
  "properties": {
    "marker-color": "#0f0"
  },
  "geometry": {
    "type": "Point",
    "coordinates": [-75.534, 39.123]
  }
}'
bearing(start, end)
}

