% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer}
\alias{buffer}
\title{Buffer a feature}
\usage{
buffer(input, dist, units = c("meters", "feet", "kilometers", "miles",
  "degrees"))
}
\arguments{
\item{input}{A Feature or FeatureCollection}

\item{dist}{distance used to buffer the input}

\item{units}{(character) Can be miles, feet, kilometers, meters, or degrees}
}
\description{
Buffer a feature
}
\details{
Calculates and returns a buffered feature.
}
\examples{
# From a FeatureCollection
dat <- random(100)
buffer(dat, 100)

# From a Feature
dat <- '{
 "type": "Feature",
 "properties": {},
 "geometry": {
     "type": "Polygon",
     "coordinates": [[
       [-112.072391,46.586591],
       [-112.072391,46.61761],
       [-112.028102,46.61761],
       [-112.028102,46.586591],
       [-112.072391,46.586591]
     ]]
   }
}'
buffer(dat, 1, "miles")
}
\author{
Jeff Hollister \email{hollister.jeff@epa.gov}
}

