% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\title{Centroid}
\usage{
centroid(features)
}
\arguments{
\item{features}{Input features}
}
\value{
Feature - centroid of the input features
}
\description{
Takes one or more features and calculates the centroid using the arithmetic
mean of all vertices. This lessens the effect of small islands and artifacts
when calculating the centroid of a set of polygons.
}
\examples{
poly <- '{
  "type": "Feature",
  "properties": {},
  "geometry": {
    "type": "Polygon",
    "coordinates": [[
      [105.818939,21.004714],
      [105.818939,21.061754],
      [105.890007,21.061754],
      [105.890007,21.004714],
      [105.818939,21.004714]
     ]]
  }
}'
centroid(poly)
}

