% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-types.R
\name{data-types}
\alias{data-Feature}
\alias{data-FeatureCollection}
\alias{data-LineString}
\alias{data-Point}
\alias{data-Polygon}
\alias{data-types}
\title{Description of data types}
\description{
Description of data types

Point description

Polygon description

LineString description

Feature description

FeatureCollection description
}

