% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/destination.R
\name{destination}
\alias{destination}
\title{Destination}
\usage{
destination(start, distance, bearing, units, lint = FALSE)
}
\arguments{
\item{start}{starting point \code{\link{data-Point}}}

\item{distance}{distance from the starting point}

\item{bearing}{ranging from -180 to 180}

\item{units}{miles, kilometers, degrees, or radians}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
destination \code{\link{data-Point}}
}
\description{
Destination
}
\details{
Takes a \code{\link{data-Point}} and calculates the location of a
destination point given a distance in degrees, radians, miles, or kilometers; and
bearing in degrees. This uses the
\href{http://en.wikipedia.org/wiki/Haversine_formula}{Haversine} formula to account
for global curvature.
}
\examples{
pt <- '{
  "type": "Feature",
  "properties": {
    "marker-color": "#0f0"
  },
  "geometry": {
    "type": "Point",
    "coordinates": [-75.343, 39.984]
  }
}'
destination(pt, 50, 90, "miles")
}

