% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Calculate the distance between two points}
\usage{
distance(from, to, units = "kilometers")
}
\arguments{
\item{polygons}{Geojson polygons}

\item{points}{Geojson points}

\item{field}{(character) The field in the points features from which to pull values to average}

\item{outputField}{(character) The field in polygons to put results of the averages}
}
\description{
Calculate the distance between two points
}
\details{
Calculates the distance between two points in degress, radians, miles, or kilometers.
Uses the Haversine formula to account for global curvature.
}
\examples{
from <- '{
 "type": "Feature",
 "properties": {},
 "geometry": {
   "type": "Point",
   "coordinates": [-75.343, 39.984]
 }
}'
to <- '{
  "type": "Feature",
  "properties": {},
  "geometry": {
    "type": "Point",
    "coordinates": [-75.534, 39.123]
  }
}'
distance(from, to)
}

