% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope}
\alias{envelope}
\title{Envelope}
\usage{
envelope(fc, lint = FALSE)
}
\arguments{
\item{fc}{\code{\link{data-FeatureCollection}}}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
a rectangular \code{\link{data-Polygon}} feature that
encompasses all vertices
}
\description{
Envelope
}
\details{
Takes any number of features and returns a rectangular
\code{\link{data-Polygon}} that encompasses all vertices.
}
\examples{
fc <- '{
  "type": "FeatureCollection",
  "features": [
    {
      "type": "Feature",
      "properties": {
        "name": "Location A"
      },
      "geometry": {
        "type": "Point",
        "coordinates": [-75.343, 39.984]
      }
    }, {
      "type": "Feature",
      "properties": {
        "name": "Location B"
      },
      "geometry": {
        "type": "Point",
        "coordinates": [-75.833, 39.284]
      }
    }, {
      "type": "Feature",
      "properties": {
        "name": "Location C"
      },
      "geometry": {
        "type": "Point",
        "coordinates": [-75.534, 39.123]
      }
    }
  ]
}'
envelope(fc)
}

