% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/explode.R
\name{explode}
\alias{explode}
\title{Explode}
\usage{
explode(input)
}
\arguments{
\item{input}{Feature of features}
}
\description{
Takes a feature or set of features and returns all positions as points
}
\examples{
poly <- '{
 "type": "Feature",
 "properties": {},
 "geometry": {
   "type": "Polygon",
   "coordinates": [[
     [177.434692, -17.77517],
     [177.402076, -17.779093],
     [177.38079, -17.803937],
     [177.40242, -17.826164],
     [177.438468, -17.824857],
     [177.454948, -17.796746],
     [177.434692, -17.77517]
    ]]
 }
}'
explode(poly)
}

