% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/turf-package.R
\docType{data}
\name{filter_features}
\alias{filter_features}
\title{Filter features}
\format{A character string - a geojson FeatureCollection of points.}
\description{
For use in \code{\link{filter}} function examples
}
\keyword{datasets}

