% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hex_grid.R
\name{hex_grid}
\alias{hex_grid}
\title{HexGrid}
\usage{
hex_grid(extent, cellWidth, units)
}
\arguments{
\item{extent}{(numeric) extent in [minX, minY, maxX, maxY] order}

\item{cellWidth}{(integer) width of each cell}

\item{units}{(character) units to use for cellWidth, one of 'miles', xxxx}
}
\value{
FeatureCollection grid of points
}
\description{
Takes a bounding box and a cell size in degrees and returns a
FeatureCollection of flat-topped hexagons (Polygon features)
aligned in an "odd-q" vertical grid as described in
Hexagonal Grids \url{http://www.redblobgames.com/grids/hexagons/}
}
\examples{
hex_grid(c(-96,31,-84,40), 50, 'miles')
hex_grid(c(-96,31,-84,40), 30, 'miles')
hex_grid(c(-96,31,-84,40), 10, 'miles')
}

