% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/isolines.R
\name{isolines}
\alias{isolines}
\title{Isolines}
\usage{
isolines(points, z, resolution, breaks)
}
\arguments{
\item{points}{Input points}

\item{z}{(character) the property name in points from which z-values will be pulled}

\item{resolution}{(numeric) resolution of the underlying grid}

\item{breaks}{(numeric) where to draw contours}
}
\value{
FeatureCollection of isolines (LineString features)
}
\description{
Takes points with z-values and an array of value breaks and generates isolines
}
\examples{
pts <- random(n = 10)
pts$features$properties <- data.frame(z = round(rnorm(10, mean = 10)), stringsAsFactors = FALSE)
breaks <- c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
isolines(pts, 'z', 15, breaks)
}

