% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jenks.R
\name{jenks}
\alias{jenks}
\title{Jenks}
\usage{
jenks(input, field, numberOfBreaks)
}
\arguments{
\item{input}{Set of Features}

\item{field}{the property in input on which to calculate Jenks natural breaks}

\item{numberOfBreaks}{Number of classes in which to group the data}
}
\value{
An array of the break values
}
\description{
Takes a set of features and returns an array of the Jenks Natural breaks for a given property
}
\examples{
jenks(lawn_data$points_average, 'population', 0)
jenks(lawn_data$points_average, 'population', 1)
jenks(lawn_data$points_average, 'population', 2)
jenks(lawn_data$points_average, 'population', 3)
jenks(lawn_data$points_average, 'population', 4)
}
\references{
Jenks Natural breaks
\url{http://en.wikipedia.org/wiki/Jenks_natural_breaks_optimization}
}

