% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-deprecated.R
\name{lawn-defunct}
\alias{lawn-defunct}
\title{Defunct functions in \pkg{lawn}}
\description{
\itemize{
\item \link{lawn_size}: Function removed. The \code{size}
method in turf.js has been removed.
See https://github.com/Turfjs/turf/issues/306
\item \link{lawn_reclass}: Function removed. The \code{reclass}
method in turf.js has been removed.
See https://github.com/Turfjs/turf/issues/306
\item \link{lawn_jenks}: Function removed. The \code{jenks}
method in turf.js has been removed.
See https://github.com/Turfjs/turf/issues/306
\item \link{lawn_quantile}: Function removed. The \code{quantile}
method in turf.js has been removed.
See https://github.com/Turfjs/turf/issues/306
\item \link{lawn_aggregate}: Function removed. The \code{aggregate}
method in turf.js has been removed.
See https://github.com/Turfjs/turf/issues/306
\item \link{lawn_idw}: Function removed.
}
}
