% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/area.R
\name{lawn_area}
\alias{lawn_area}
\title{Calculate the area of a polygon or group of polygons}
\usage{
lawn_area(input, lint = FALSE)
}
\arguments{
\item{input}{A feature or feature collection of polygons.}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\description{
Calculate the area of a polygon or group of polygons
}
\details{
Calculates the total area of a polygon or group of polygons. Returns
         the values in square meters.
}
\examples{
lawn_area(lawn_data$poly)
lawn_area(lawn_data$multipoly)
}

