% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/average.R
\name{lawn_average}
\alias{lawn_average}
\title{Calculate the average value of a field for a set of points within a set of polygons}
\usage{
lawn_average(polygons, points, field, outputField = "average", lint = FALSE)
}
\arguments{
\item{polygons}{Geojson polygons}

\item{points}{Geojson points}

\item{field}{(character) The field in the points features from which to pull values to average}

\item{outputField}{(character) The field in polygons to put results of the averages}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
polygons with the value of outField set to the calculated averages
}
\description{
Calculate the average value of a field for a set of points within a set of polygons
}
\examples{
# using data in the package
cat(lawn_data$points_average)
cat(lawn_data$polygons_average)
lawn_average(polygons = lawn_data$polygons_average, points = lawn_data$points_average, 'population')
}

