% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bbox_polygon.R
\name{lawn_bbox_polygon}
\alias{lawn_bbox_polygon}
\title{Bbox polygon}
\usage{
lawn_bbox_polygon(bbox)
}
\arguments{
\item{bbox}{an Array of bounding box coordinates in the form:
[xLow, yLow, xHigh, yHigh]}
}
\value{
a \code{\link{data-Polygon}} representation of the bounding box
}
\description{
Bbox polygon
}
\details{
Takes a bbox and returns an equivalent polygon \code{\link{data-Polygon}}
}
\examples{
bbox <- c(0, 0, 10, 10)
lawn_bbox_polygon(bbox)
}

