% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/center.R
\name{lawn_center}
\alias{lawn_center}
\title{Center}
\usage{
lawn_center(features, lint = FALSE)
}
\arguments{
\item{features}{input featurs, as a \code{\link{data-FeatureCollection}}}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
a \code{\link{data-Point}} feature at the absolute center point of
all input features
}
\description{
Center
}
\details{
Takes a \code{\link{data-FeatureCollection}} and returns the
absolute center point of all features.
}
\examples{
lawn_center(lawn_data$points_average)
}

