% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/centroid.R
\name{lawn_centroid}
\alias{lawn_centroid}
\title{Centroid}
\usage{
lawn_centroid(features, lint = FALSE)
}
\arguments{
\item{features}{Input features}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
Feature - centroid of the input features
}
\description{
Takes one or more features and calculates the centroid using the arithmetic
mean of all vertices. This lessens the effect of small islands and artifacts
when calculating the centroid of a set of polygons.
}
\examples{
poly <- '{
  "type": "Feature",
  "properties": {},
  "geometry": {
    "type": "Polygon",
    "coordinates": [[
      [105.818939,21.004714],
      [105.818939,21.061754],
      [105.890007,21.061754],
      [105.890007,21.004714],
      [105.818939,21.004714]
     ]]
  }
}'
lawn_centroid(poly)
}

