% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distance.R
\name{lawn_distance}
\alias{lawn_distance}
\title{Calculate the distance between two points}
\usage{
lawn_distance(from, to, units = "kilometers", lint = FALSE)
}
\arguments{
\item{from}{Origin point}

\item{to}{Destination point}

\item{units}{(character) Can be degrees, radians, miles, or kilometers}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\description{
Calculate the distance between two points
}
\details{
Calculates the distance between two points in degress, radians, miles, or kilometers.
Uses the Haversine formula to account for global curvature.
}
\examples{
from <- '{
 "type": "Feature",
 "properties": {},
 "geometry": {
   "type": "Point",
   "coordinates": [-75.343, 39.984]
 }
}'
to <- '{
  "type": "Feature",
  "properties": {},
  "geometry": {
    "type": "Point",
    "coordinates": [-75.534, 39.123]
  }
}'
lawn_distance(from, to)
}

