% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.R
\name{lawn_filter}
\alias{lawn_filter}
\title{Filter a FeatureCollection by a given property and value}
\usage{
lawn_filter(features, key, value)
}
\arguments{
\item{features}{A FeatureCollection}

\item{key}{(character) The property on which to filter}

\item{value}{(character) The value of that property on which to filter}
}
\description{
Filter a FeatureCollection by a given property and value
}
\examples{
cat(lawn_data$filter_features)
lawn_filter(lawn_data$filter_features, 'species', 'oak')
lawn_filter(lawn_data$filter_features, 'species', 'maple')
lawn_filter(lawn_data$filter_features, 'species', 'redwood')
}

