% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{lawn_flatten}
\alias{lawn_flatten}
\title{Flatten}
\usage{
lawn_flatten(x, lint = FALSE)
}
\arguments{
\item{x}{any valid GeoJSON with multi-geometry Feature's}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
a \link{data-FeatureCollection}
}
\description{
Flattens any GeoJSON to a FeatureCollection
}
\examples{
x <- '{
"type": "MultiPolygon",
"coordinates": [
    [[[102.0, 2.0], [103.0, 2.0], [103.0, 3.0], [102.0, 3.0], [102.0, 2.0]]],
    [[[100.0, 0.0], [101.0, 0.0], [101.0, 1.0], [100.0, 1.0], [100.0, 0.0]],
    [[100.2, 0.2], [100.8, 0.2], [100.8, 0.8], [100.2, 0.8], [100.2, 0.2]]]
  ]
}'
lawn_flatten(x)
}
\seealso{
Other misc: \code{\link{lawn_truncate}}
}
