% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{lawn_flip}
\alias{lawn_flip}
\title{Flip x,y to y,x, and vice versa}
\usage{
lawn_flip(input, lint = FALSE)
}
\arguments{
\item{input}{Feature of features.}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
A \link{data-Feature} or \link{data-FeatureCollection}.
}
\description{
Flip x,y to y,x, and vice versa
}
\examples{
# a point
serbia <- '{
  "type": "Feature",
  "properties": {"color": "red"},
  "geometry": {
    "type": "Point",
    "coordinates": [20.566406, 43.421008]
   }
}'
lawn_flip(serbia)

# a featurecollection
pts <- lawn_random("points")
lawn_flip(pts)
\dontrun{
lawn_data$points_average \%>\% view
lawn_flip(lawn_data$points_average) \%>\% view
lawn_data$polygons_average \%>\% view
lawn_flip(lawn_data$polygons_average) \%>\% view
}
}
