% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrycollection.R
\name{lawn_geometrycollection}
\alias{lawn_geometrycollection}
\title{Create a geometrycollection}
\usage{
lawn_geometrycollection(coordinates, properties = NULL)
}
\arguments{
\item{coordinates}{A list of GeoJSON geometries, or in json}

\item{properties}{A list of properties}
}
\value{
a \code{\link{data-GeometryCollection}} feature
}
\description{
Create a geometrycollection
}
\seealso{
Other data functions: \code{\link{lawn_featurecollection}},
  \code{\link{lawn_feature}}, \code{\link{lawn_filter}},
  \code{\link{lawn_linestring}},
  \code{\link{lawn_multilinestring}},
  \code{\link{lawn_multipoint}},
  \code{\link{lawn_multipolygon}},
  \code{\link{lawn_point}}, \code{\link{lawn_polygon}},
  \code{\link{lawn_random}}, \code{\link{lawn_remove}},
  \code{\link{lawn_sample}}
}

