% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinks.R
\name{lawn_kinks}
\alias{lawn_kinks}
\title{Get points at all self-intersections of a polygon}
\usage{
lawn_kinks(input, lint = FALSE)
}
\arguments{
\item{input}{Feature of features.}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing
time as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good
geojson objects. Default: \code{FALSE}}
}
\description{
Get points at all self-intersections of a polygon
}
\examples{
poly <- '{
 "type": "Feature",
 "properties": {},
 "geometry": {
   "type": "Polygon",
   "coordinates": [[
      [-12.034835, 8.901183],
      [-12.060413, 8.899826],
      [-12.03638, 8.873199],
      [-12.059383, 8.871418],
      [-12.034835, 8.901183]
    ]]
  }
}'
lawn_kinks(poly)
# lint input object
# lawn_kinks(poly, TRUE)
\dontrun{
poly \%>\% view
lawn_kinks(poly) \%>\% view
}
}
