% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge.R
\name{lawn_merge}
\alias{lawn_merge}
\title{Merge}
\usage{
lawn_merge(fc, lint = FALSE)
}
\arguments{
\item{fc}{input polygons, as \code{\link{data-FeatureCollection}}}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
merged \code{\link{data-Polygon}} or multipolygon
\code{\link{data-MultiPolygon}}
}
\description{
Merge
}
\details{
Takes a set of \code{\link{data-Polygon}} and returns a single
merged polygon feature. If the input polygon features are not contiguous,
this function returns a \code{\link{data-MultiPolygon}} feature.
}
\examples{
polygons <- '{
  "type": "FeatureCollection",
  "features": [
    {
      "type": "Feature",
      "properties": {
        "fill": "#0f0"
      },
      "geometry": {
        "type": "Polygon",
        "coordinates": [[
          [9.994812, 53.549487],
          [10.046997, 53.598209],
          [10.117721, 53.531737],
          [9.994812, 53.549487]
        ]]
      }
    }, {
      "type": "Feature",
      "properties": {
        "fill": "#00f"
      },
      "geometry": {
        "type": "Polygon",
        "coordinates": [[
          [10.000991, 53.50418],
          [10.03807, 53.562539],
          [9.926834, 53.551731],
          [10.000991, 53.50418]
        ]]
      }
    }
  ]
}'
lawn_merge(polygons)
}

