% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/point.R
\name{lawn_point}
\alias{lawn_point}
\title{Create a point}
\usage{
lawn_point(x, properties = NULL)
}
\arguments{
\item{x}{A pair of points in a vector, list or json, of the form
\code{e.g., c(longitude,latitude)}}

\item{properties}{A list of properties. Default: NULL}
}
\description{
Create a point
}
\examples{
lawn_point(c(-74.5, 40))
lawn_point(list(-74.5, 40))
lawn_point('[-74.5, 40]')
lawn_point(c(-74.5, 40), properties = list(name = 'poly1', population = 400))

# Make a FeatureCollection
# lawn_featurecollection(lawn_point(c(-74.5, 40)))
}

