% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/size.R
\name{lawn_size}
\alias{lawn_size}
\title{Size}
\usage{
lawn_size(bbox, factor)
}
\arguments{
\item{bbox}{a bounding box}

\item{factor}{the ratio of the new bbox to the input bbox}
}
\value{
the resized bbox
}
\description{
Size
}
\details{
Takes a bounding box and returns a new bounding box with a size
expanded or contracted by a factor of X.
}
\examples{
bbox <- c(0, 0, 10, 10)
lawn_size(bbox, factor = 1)
lawn_size(bbox, factor = 10)
lawn_size(bbox, factor = 50)
lawn_size(bbox, factor = 100)
}

