% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/square.R
\name{lawn_square}
\alias{lawn_square}
\title{Square}
\usage{
lawn_square(bbox, lint = FALSE)
}
\arguments{
\item{bbox}{A bounding box}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
A square surrounding bbox
}
\description{
Takes a bounding box and calculates the minimum square bounding
box that would contain the input.
}
\examples{
bbox <- c(-20, -20, -15, 0)
lawn_square(bbox)
lawn_square(bbox, TRUE)
sq <- lawn_square(bbox)
}
\seealso{
Other measurements: \code{\link{lawn_along}};
  \code{\link{lawn_area}}; \code{\link{lawn_bbox_polygon}};
  \code{\link{lawn_bearing}}; \code{\link{lawn_center}};
  \code{\link{lawn_centroid}};
  \code{\link{lawn_destination}};
  \code{\link{lawn_distance}}; \code{\link{lawn_envelope}};
  \code{\link{lawn_extent}};
  \code{\link{lawn_line_distance}};
  \code{\link{lawn_midpoint}};
  \code{\link{lawn_point_on_surface}};
  \code{\link{lawn_size}}
}

