% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/line_distance.R
\name{line_distance}
\alias{line_distance}
\title{Line distance}
\usage{
line_distance(line, units, lint = FALSE)
}
\arguments{
\item{line}{Line to measure, a \code{\link{data-LineString}}}

\item{units}{Can be degrees, radians, miles, or kilometers}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
length of the input line
}
\description{
Line distance
}
\details{
Takes a \code{\link{data-LineString}} and measures its length in
the specified units.
}
\examples{
line <- '{
  "type": "Feature",
  "properties": {},
  "geometry": {
    "type": "LineString",
    "coordinates": [
      [-77.031669, 38.878605],
      [-77.029609, 38.881946],
      [-77.020339, 38.884084],
      [-77.025661, 38.885821],
      [-77.021884, 38.889563],
      [-77.019824, 38.892368]
    ]
  }
}'
line_distance(line, 'kilometers')
line_distance(line, 'miles')
line_distance(line, 'radians')
line_distance(line, 'degrees')
}

