% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/linestring.R
\name{linestring}
\alias{linestring}
\title{Create a linestring}
\usage{
linestring(coordinates, properties = NULL)
}
\arguments{
\item{coordinates}{A list of Positions}

\item{properties}{A list of properties}
}
\description{
Create a linestring
}
\examples{
pts <- list(
   c(-21.964416, 64.148203),
   c(-21.956176, 64.141316),
   c(-21.93901, 64.135924),
   c(-21.927337, 64.136673)
)
linestring(pts, properties = list(name = 'line1', distance = 145))

# completely non-sensical, but gets some data quickly
pts <- random()$features$geometry$coordinates
linestring(pts)
}

