% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/midpoint.R
\name{midpoint}
\alias{midpoint}
\title{Midpoint}
\usage{
midpoint(pt1, pt2, lint = FALSE)
}
\arguments{
\item{pt1}{First point}

\item{pt2}{Second point}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
A \code{\link{data-Point}} midway between pt1 and pt2
}
\description{
Midpoint
}
\details{
Takes two \code{\link{data-Point}} and returns a point midway
between them
}
\examples{
pt1 <- '{
  "type": "Feature",
  "properties": {},
  "geometry": {
    "type": "Point",
    "coordinates": [144.834823, -37.771257]
  }
}'
pt2 <- '{
  "type": "Feature",
  "properties": {},
  "geometry": {
    "type": "Point",
    "coordinates": [145.14244, -37.830937]
  }
}'
midpoint(pt1, pt2)
}

