% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\title{Create a point}
\usage{
point(x, properties = NULL)
}
\arguments{
\item{x}{A pair of points, of the form \code{c(longitude,latitude)}}

\item{properties}{A list of properties}
}
\description{
Create a point
}
\examples{
point(c(40, -74.50))
point(c(40, -74.50), properties = list(name = 'poly1', population = 400))
}

