% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/point_grid.R
\name{point_grid}
\alias{point_grid}
\title{PointGrid}
\usage{
point_grid(extent, cellWidth, units)
}
\arguments{
\item{extent}{(numeric) extent in [minX, minY, maxX, maxY] order}

\item{cellWidth}{(integer) width of each cell}

\item{units}{(character) units to use for cellWidth, one of 'miles', xxxx}
}
\value{
FeatureCollection grid of points
}
\description{
Takes a bounding box and a cell depth and returns a set of points in a grid
}
\examples{
point_grid(c(-77.3876, 38.7198, -76.9482, 39.0277), 30, 'miles')
point_grid(c(-77.3876, 38.7198, -76.9482, 39.0277), 10, 'miles')
point_grid(c(-77.3876, 38.7198, -76.9482, 39.0277), 3, 'miles')
}

