% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/turf-package.R
\docType{data}
\name{points_average}
\alias{points_average}
\title{Average points}
\format{A character string - a geojson FeatureCollection of points.}
\description{
For use in \code{\link{average}} function examples
}
\keyword{datasets}

