% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/turf-package.R
\docType{data}
\name{points_count}
\alias{points_count}
\title{Count points}
\format{A character string - a geojson FeatureCollection of points.}
\description{
For use in \code{\link{count}} function examples
}
\keyword{datasets}

