% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polygon.R
\name{polygon}
\alias{polygon}
\title{Create a polygon}
\usage{
polygon(rings, properties = NULL)
}
\arguments{
\item{rings}{A list of LinearRings}

\item{properties}{A list of properties}
}
\description{
Create a polygon
}
\examples{
rings <- list(list(
   c(-2.275543, 53.464547),
   c(-2.275543, 53.489271),
   c(-2.215118, 53.489271),
   c(-2.215118, 53.464547),
   c(-2.275543, 53.464547)
))
polygon(rings, properties = list(name = 'poly1', population = 400))

# Make a FeatureCollection
featurecollection(polygon(rings))
}

