% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lawn-package.R
\docType{data}
\name{polygons_average}
\alias{polygons_average}
\title{Average polygons}
\format{A character string - a geojson FeatureCollection of two polygons}
\description{
For use in \code{\link{average}} function examples
}
\keyword{datasets}

