% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lawn-package.R
\docType{data}
\name{polygons_count}
\alias{polygons_count}
\title{Count polygons}
\format{A character string - a geojson FeatureCollection of polygons.}
\description{
For use in \code{\link{count}} function examples
}
\keyword{datasets}

