% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print-methods}
\alias{print-methods}
\title{Lawn print methods to provide summary view}
\arguments{
\item{x}{input}

\item{...}{print options}
}
\description{
Lawn print methods to provide summary view
}
\examples{
# point
lawn_point(c(-74.5, 40))

# polygon
rings <- list(list(
   c(-2.275543, 53.464547),
   c(-2.275543, 53.489271),
   c(-2.215118, 53.489271),
   c(-2.215118, 53.464547),
   c(-2.275543, 53.464547)
))
lawn_polygon(rings, properties = list(name = 'poly1', population = 400))

# linestring
linestring1 <- '[
   [-21.964416, 64.148203],
   [-21.956176, 64.141316],
   [-21.93901, 64.135924],
   [-21.927337, 64.136673]
]'
lawn_linestring(linestring1)
lawn_linestring(linestring1, properties = list(name = 'line1', distance = 145))

# featurecollection
lawn_featurecollection(lawn_data$featurecollection_eg1)
}

