% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile}
\alias{quantile}
\title{Quantile}
\usage{
quantile(input, field, percentiles)
}
\arguments{
\item{input}{Set of Features}

\item{field}{The property in input from which to retrieve quantile values}

\item{percentiles}{An Array of percentiles on which to calculate quantile values}
}
\value{
An array of the break values
}
\description{
Takes a FeatureCollection, a property name, set of percentiles and returns a quantile array
}
\examples{
pts <- '{
 "type": "FeatureCollection",
 "features": [
   {
     "type": "Feature",
     "properties": {
       "population": 5
     },
     "geometry": {
       "type": "Point",
       "coordinates": [5, 5]
     }
   }, {
     "type": "Feature",
     "properties": {
       "population": 40
     },
     "geometry": {
       "type": "Point",
       "coordinates": [1, 3]
     }
   }, {
     "type": "Feature",
     "properties": {
       "population": 80
     },
     "geometry": {
       "type": "Point",
       "coordinates": [14, 2]
     }
   }, {
     "type": "Feature",
     "properties": {
       "population": 90
     },
     "geometry": {
       "type": "Point",
       "coordinates": [13, 1]
     }
   }, {
     "type": "Feature",
     "properties": {
       "population": 100
     },
     "geometry": {
       "type": "Point",
       "coordinates": [19, 7]
     }
   }
 ]
}'
quantile(pts, 'population', c(25, 50, 75, 99))
quantile(pts, 'population', c(25, 50, 75, 95))
quantile(pts, 'population', c(25, 50, 75))
quantile(pts, 'population', c(1, 50, 99.9))
}

