% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample.R
\name{random}
\alias{random}
\alias{sample}
\title{Return a FeatureCollection with given number of features at random}
\usage{
random(type = "points", n = 10, bbox = NULL, num_vertices = NULL,
  max_radial_length = NULL)

sample(features = NULL, n = 100)
}
\arguments{
\item{type}{type of features desired: 'points' or 'polygons'}

\item{n}{(integer) Number of features to generate}

\item{features}{A FeatureCollection}

\item{options}{options relevant to the feature desired. See Details.}
}
\description{
Return a FeatureCollection with given number of features at random
}
\details{
Options fore the \code{options} parameter:
\itemize{
 \item \code{bbox} A bounding box inside of which geometries are placed. In the case of Point features, they are guaranteed to be within this bounds, while Polygon features have their centroid within the bounds.
 \item \code{num_vertices}	Number options.vertices the number of vertices added to polygon features.
 \item \code{max_radial_length} Number	<optional> 10	 the total number of decimal degrees longitude or latitude that a polygon can extent outwards to from its center.
 }
}
\examples{
# Random
## set of points
random(2)
random(10)
random(1000)
## set of polygons
random('polygons', 2)
random('polygons', 10)
random('polygons', 1000)
# with options
random(bbox = c(-70, 40, -60, 60))
random(num_vertices = 5)

# Sample points from a FeatureCollection
sample(lawn_data$points_average, 1)
sample(lawn_data$points_average, 2)
sample(lawn_data$points_average, 3)
}

