% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remove.R
\name{remove}
\alias{remove}
\title{Return a FeatureCollection with features removed matching value of property given}
\usage{
remove(features, property, value)
}
\arguments{
\item{features}{A set of input features}

\item{property}{Property to filter}

\item{value}{Value to filter}
}
\value{
A FeatureCollection
}
\description{
Return a FeatureCollection with features removed matching value of property given
}
\examples{
cat(lawn_data$remove_features)
remove(lawn_data$remove_features, 'marker-color', '#00f')
remove(lawn_data$remove_features, 'marker-color', '#0f0')
}

