% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplify}
\usage{
simplify(feature, tolerance, high_quality = FALSE, lint = FALSE)
}
\arguments{
\item{feature}{a \code{\link{data-LineString}} or \code{\link{data-Polygon}}
feature to be simplified}

\item{tolerance}{simplification tolerance}

\item{high_quality}{(boolean) whether or not to spend more time to create a
higher-quality simplification with a different algorithm}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
a simplified feature

A FeatureCollection of \code{\link{data-Polygon}} features with
properties listed as \code{out_field}
}
\description{
Simplify
}
\details{
Takes a \code{\link{data-LineString}} or \code{\link{data-Polygon}}
feature and returns a simplified version. Internally uses simplify-js
(\url{http://mourner.github.io/simplify-js/}) to perform simplification.
}
\examples{
feature <- '{
  "type": "Feature",
  "properties": {},
  "geometry": {
    "type": "Polygon",
    "coordinates": [[
      [-70.603637, -33.399918],
      [-70.614624, -33.395332],
      [-70.639343, -33.392466],
      [-70.659942, -33.394759],
      [-70.683975, -33.404504],
      [-70.697021, -33.419406],
      [-70.701141, -33.434306],
      [-70.700454, -33.446339],
      [-70.694274, -33.458369],
      [-70.682601, -33.465816],
      [-70.668869, -33.472117],
      [-70.646209, -33.473835],
      [-70.624923, -33.472117],
      [-70.609817, -33.468107],
      [-70.595397, -33.458369],
      [-70.587158, -33.442901],
      [-70.587158, -33.426283],
      [-70.590591, -33.414248],
      [-70.594711, -33.406224],
      [-70.603637, -33.399918]
    ]]
  }
}'

simplify(feature, tolerance = 0.01)
}

