% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\title{Inside}
\usage{
tag(points, polygons, polyId, caontainingPolyId)
}
\arguments{
\item{point}{Input point}

\item{polygon}{Input polygon or multipolygon}
}
\value{
\code{TRUE} if the Point IS inside the Polygon, \code{FALSE} if
the Point IS NOT inside the Polygon
}
\description{
Takes a Point and a Polygon or MultiPolygon and determines if the point
resides inside the polygon. The polygon can be convex or concave. The
function accounts for holes
}
\examples{
pts <- random(30)
polys <- triangle_grid(c(-77.3876, 38.7198, -76.9482, 39.0277), 30, 'miles')
tag(pts, polys, 'fill', 'marker-color')
}

