% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/turf-package.R
\docType{package}
\name{turf-package}
\alias{turf}
\alias{turf-package}
\title{R client for turf.js for geospatial analysis}
\description{
R client for turf.js for geospatial analysis
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

