% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{Visualize}
\usage{
view(x)
}
\arguments{
\item{x}{Input, a geojson character string}
}
\description{
Visualize
}
\examples{
\dontrun{
view(lawn_data$polygons_average)
view(lawn_data$filter_features)
view(lawn_data$polygons_within)
view(lawn_data$polygons_count)
}
}

