% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\title{Calculate the area of a polygon or group of polygons}
\usage{
area(input)
}
\arguments{
\item{input}{A feature or feature collection of polygons.}
}
\description{
Calculate the area of a polygon or group of polygons
}
\details{
Calculates the total area of a polygon or group of polygons. Returns
         the values in square meters.
}
\examples{
area(lawn_data$poly)
area(lawn_data$multipoly)
}

