% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/average.R
\name{average}
\alias{average}
\title{Calculate the average value of a field for a set of points within a set of polygons.}
\usage{
average(polygons, points, field, outputField = "average")
}
\arguments{
\item{polygons}{Geojson polygons}

\item{points}{Geojson points}

\item{field}{(character) The field in the points features from which to pull values to average}

\item{outputField}{(character) The field in polygons to put results of the averages}
}
\description{
Calculate the average value of a field for a set of points within a set of polygons.
}
\examples{
# using data in the package
cat(lawn_data$points_average)
cat(lawn_data$polygons_average)
average(polygons = lawn_data$polygons_average, points = lawn_data$points_average, 'population')
}

