% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Calculates the number of points that fall within the set of polygons.}
\usage{
count(polygons, points, countField = "pt_count")
}
\arguments{
\item{polygons}{Geojson polygons}

\item{points}{Geojson points}

\item{countField}{(character) A field to append to the attributes of the Polygon features
representing Point counts}
}
\description{
Calculates the number of points that fall within the set of polygons.
}
\examples{
# using data in the package
count(polygons = polygons_count, points = points_count)
}

