% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/count.R
\name{count}
\alias{count}
\title{Calculates the number of points that fall within the set of polygons.}
\usage{
count(polygons, points, countField = "pt_count")
}
\arguments{
\item{polygons}{Geojson polygons}

\item{points}{Geojson points}

\item{countField}{(character) A field to append to the attributes of the Polygon features
representing Point counts}
}
\description{
Calculates the number of points that fall within the set of polygons.
}
\examples{
polygons <- '{
  "type": "FeatureCollection",
  "features": [
    {
      "type": "Feature",
      "properties": {},
      "geometry": {
        "type": "Polygon",
        "coordinates": [[
          [-112.072391,46.586591],
          [-112.072391,46.61761],
          [-112.028102,46.61761],
          [-112.028102,46.586591],
          [-112.072391,46.586591]
          ]]
      }
    }, {
      "type": "Feature",
      "properties": {},
      "geometry": {
        "type": "Polygon",
        "coordinates": [[
          [-112.023983,46.570426],
          [-112.023983,46.615016],
          [-111.966133,46.615016],
          [-111.966133,46.570426],
          [-112.023983,46.570426]
          ]]
      }
    }
    ]
};'
points <- '{
"type": "FeatureCollection",
"features": [
    {
      "type": "Feature",
      "properties": {
        "population": 200
      },
      "geometry": {
        "type": "Point",
        "coordinates": [-112.0372, 46.608058]
      }
    }, {
      "type": "Feature",
      "properties": {
        "population": 600
      },
      "geometry": {
        "type": "Point",
        "coordinates": [-112.045955, 46.596264]
      }
    }
    ]
};'
count(polygons = polygons, points = points)
}

