% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count}
\alias{count}
\title{Calculates the number of points that fall within the set of polygons.}
\usage{
count(polygons, points, countField = "pt_count")
}
\arguments{
\item{polygons}{Geojson polygons}

\item{points}{Geojson points}

\item{countField}{(character) A field to append to the attributes of the Polygon features
representing Point counts}
}
\description{
Calculates the number of points that fall within the set of polygons.
}
\examples{
# using data in the package
cat(lawn_data$points_count)
cat(lawn_data$polygons_count)
count(polygons = lawn_data$polygons_count, points = lawn_data$points_count)
}

