% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-types.R
\name{data-types}
\alias{data-Feature}
\alias{data-FeatureCollection}
\alias{data-LineString}
\alias{data-Point}
\alias{data-Polygon}
\alias{data-types}
\title{Description of data types}
\description{
Description of data types
}
\section{Point}{


Point is a xyz
}

\section{Polygon}{


Polygon is a xyz
}

\section{LineString}{


LineString is a xyz
}

\section{Feature}{


Feature is a xyz
}

\section{FeatureCollection}{


FeatureCollection is a xyz
}

