% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter}
\alias{filter}
\title{Filter a FeatureCollection by a given property and value}
\usage{
filter(features, key, value)
}
\arguments{
\item{features}{A FeatureCollection}

\item{key}{(character) The property on which to filter}

\item{value}{(character) The value of that property on which to filter}
}
\description{
Filter a FeatureCollection by a given property and value
}
\examples{
cat(lawn_data$filter_features)
filter(lawn_data$filter_features, 'species', 'oak')
filter(lawn_data$filter_features, 'species', 'maple')
filter(lawn_data$filter_features, 'species', 'redwood')
}

