% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Filter a FeatureCollection by a given property and value}
\usage{
filter(features, key, value)
}
\arguments{
\item{features}{A FeatureCollection}

\item{key}{(character) The property on which to filter}

\item{value}{(character) The value of that property on which to filter}
}
\description{
Filter a FeatureCollection by a given property and value
}
\examples{
dat <- filter_features
filter(dat, 'species', 'oak')
filter(dat, 'species', 'maple')
filter(dat, 'species', 'redwood')
}

