% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip x,y to y,x, and vice versa}
\usage{
flip(input, lint = FALSE)
}
\arguments{
\item{input}{Feature of features}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\description{
Flip x,y to y,x, and vice versa
}
\examples{
serbia <- '{
  "type": "Feature",
  "properties": {},
  "geometry": {
    "type": "Point",
    "coordinates": [20.566406, 43.421008]
   }
}'
flip(serbia)
}

