% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip x,y to y,x, and vice versa}
\usage{
flip(input, lint = FALSE)
}
\arguments{
\item{input}{Feature of features}
}
\description{
Flip x,y to y,x, and vice versa
}
\examples{
serbia <- '{
  "type": "Feature",
  "properties": {},
  "geometry": {
    "type": "Point",
    "coordinates": [20.566406, 43.421008]
   }
}'
flip(serbia)
}

