% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kinks.R
\name{kinks}
\alias{kinks}
\title{Kinks}
\usage{
kinks(input)
}
\arguments{
\item{input}{Feature of features}
}
\description{
Takes a polygon and returns points at all self-intersections
}
\examples{
poly <- '{
 "type": "Feature",
 "properties": {},
 "geometry": {
   "type": "Polygon",
   "coordinates": [[
     [-12.034835, 8.901183],
     [-12.060413, 8.899826],
     [-12.03638, 8.873199],
     [-12.059383, 8.871418],
     [-12.034835, 8.901183]
    ]]
  }
}'
kinks(poly)
}

