% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-deprecated.R
\name{lawn-defunct}
\alias{lawn-defunct}
\title{Defunct functions in \pkg{lawn}}
\description{
\itemize{
 \item \code{\link{lawn_size}}: Function removed. The \code{size}
 method in turf.js has been removed.
 See \url{https://github.com/Turfjs/turf/issues/306}
 \item \code{\link{lawn_reclass}}: Function removed. The \code{reclass}
 method in turf.js has been removed.
 See \url{https://github.com/Turfjs/turf/issues/306}
}
}

