% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lawn-package.R
\docType{package}
\name{lawn-package}
\alias{lawn-package}
\alias{lawn}
\title{R client for turf.js for geospatial analysis}
\description{
turf.js uses GeoJSON for all geographic data, and expects the data to be
standard \href{http://en.wikipedia.org/wiki/World_Geodetic_System}{WGS84}
longitude,latitude coordinates. See https://geojson.io/ for a tool to
easily create GeoJSON in a browser.
}
\seealso{
\link{lawn-defunct}
}
\author{
Scott Chamberlain

Jeff Hollister (\email{hollister.jeff@epa.gov})
}
