% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-deprecated.R
\name{lawn_aggregate}
\alias{lawn_aggregate}
\title{Calculate a series of aggregations for a set of points within polygons}
\usage{
lawn_aggregate(...)
}
\description{
This function is defunct
}
\keyword{internal}

