% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/area.R
\name{lawn_area}
\alias{lawn_area}
\title{Calculate the area of a polygon or group of polygons}
\usage{
lawn_area(input, lint = FALSE)
}
\arguments{
\item{input}{A feature or featurecollection of polygons.}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
value in square meters
}
\description{
Calculate the area of a polygon or group of polygons
}
\examples{
lawn_area(lawn_data$poly)
lawn_area(lawn_data$multipoly)
}
\seealso{
Other measurements: \code{\link{lawn_along}};
  \code{\link{lawn_bbox_polygon}};
  \code{\link{lawn_bearing}}; \code{\link{lawn_center}};
  \code{\link{lawn_centroid}};
  \code{\link{lawn_destination}};
  \code{\link{lawn_distance}}; \code{\link{lawn_envelope}};
  \code{\link{lawn_extent}};
  \code{\link{lawn_line_distance}};
  \code{\link{lawn_midpoint}};
  \code{\link{lawn_point_on_surface}};
  \code{\link{lawn_size}}; \code{\link{lawn_square}}
}

