% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deviation.R
\name{lawn_deviation}
\alias{lawn_deviation}
\title{Deviation}
\usage{
lawn_deviation(polys, pts, inField, outField, lint = FALSE)
}
\arguments{
\item{polys}{Polygon(s) defining area to aggregate.}

\item{pts}{Points with values to aggregate.}

\item{inField}{Characater for the name of the field on \code{pts} on which you wish to perfrom
the aggregation}

\item{outField}{Characater for the name of the field on the ouput polygon FeatureCollection
that will store the resultant value.}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\description{
Deviation
}
\details{
Calculates the population standard deviation (i.e. denominator = n, not n-1) of
         values from points that are within a polygon and returns that on a polygon
         FeatureCollection.
}
\examples{
ex_polys <- lawn_data$polygons_aggregate
ex_pts <- lawn_data$points_aggregate
lawn_deviation(ex_polys, ex_pts, "population", "pop_sd")
}
\author{
Jeff Hollister \email{hollister.jeff@epa.gov}
}

