% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{lawn_flatten}
\alias{lawn_flatten}
\title{Flatten}
\usage{
lawn_flatten(x, lint = FALSE)
}
\arguments{
\item{x}{any valid GeoJSON with multi-geometry \link{data-Feature}'s}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing
time as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good
geojson objects. Default: \code{FALSE}}
}
\value{
a \link{data-FeatureCollection}
}
\description{
Flattens any GeoJSON to a FeatureCollection
}
\examples{
x <- '{"type":"MultiPolygon","coordinates":[
  [[[102,2],[103,2],[103,3],[102,3],[102,2]]],
  [[[100,0],[101,0],[101,1],[100,1],[100,0]],
  [[100.2,0.2],[100.2,0.8],[100.8,0.8],[100.8,0.2],[100.2,0.2]]]
 ]
}'
lawn_flatten(x)
lawn_flatten(x, TRUE)
}
\seealso{
Other misc: \code{\link{lawn_truncate}}
}
