% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/square_grid.R
\name{lawn_square_grid}
\alias{lawn_square_grid}
\title{SquareGrid}
\usage{
lawn_square_grid(extent, cellWidth, units)
}
\arguments{
\item{extent}{(numeric) extent in [minX, minY, maxX, maxY] order}

\item{cellWidth}{(integer) width of each cell}

\item{units}{(character) units to use for cellWidth, one of 'miles', xxxx}
}
\value{
FeatureCollection grid of polygons
}
\description{
Takes a bounding box and a cell depth and returns a set of square polygons in a grid
}
\examples{
lawn_square_grid(c(-77.3876, 38.7198, -76.9482, 39.0277), 30, 'miles')
lawn_square_grid(c(-77.3876, 38.7198, -76.9482, 39.0277), 10, 'miles')
lawn_square_grid(c(-77.3876, 38.7198, -76.9482, 39.0277), 3, 'miles')
}

