% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tag.R
\name{lawn_tag}
\alias{lawn_tag}
\title{Tag}
\usage{
lawn_tag(points, polygons, polyId, containingPolyId, lint = FALSE)
}
\arguments{
\item{points}{Input \code{\link{data-Point}}}

\item{polygons}{Input \code{\link{data-Polygon}} or \code{\link{data-MultiPolygon}}}

\item{polyId}{property in polygons to add to joined Point features}

\item{containingPolyId}{property in points in which to store joined
property from polygons}

\item{lint}{(logical) Lint or not. Uses geojsonhint. Takes up increasing time
as the object to get linted increases in size, so probably use by
default for small objects, but not for large if you know they are good geojson
objects. Default: \code{FALSE}}
}
\value{
points with containingPolyId property containing values from polyId
}
\description{
Tag
}
\details{
Takes a set of points and a set of polygons and performs a spatial join
}
\examples{
pts <- lawn_random(n = 30)
polys <- lawn_triangle_grid(c(-77.3876, 38.7198, -76.9482, 39.0277), 30, 'miles')
lawn_tag(pts, polys, 'fill', 'marker-color')
}

