% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/within.R
\name{lawn_within}
\alias{lawn_within}
\title{Return points that fall within polygons}
\usage{
lawn_within(points, polygons)
}
\arguments{
\item{points}{FeatureCollection of points}

\item{polygons}{FeatureCollection of polygons}
}
\description{
Return points that fall within polygons
}
\examples{
cat(lawn_data$points_within)
cat(lawn_data$polygons_within)
lawn_within(lawn_data$points_within, lawn_data$polygons_within)
}

