% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\title{Create a point}
\usage{
point(x, properties = NULL)
}
\arguments{
\item{x}{A pair of points in a vector, list or json, of the form
\code{e.g., c(longitude,latitude)}}

\item{properties}{A list of properties. Default: NULL}
}
\description{
Create a point
}
\examples{
point(c(-74.5, 40))
point(list(-74.5, 40))
point('[-74.5, 40]')
point(c(-74.5, 40), properties = list(name = 'poly1', population = 400))

# Make a FeatureCollection
featurecollection(point(c(-74.5, 40)))
}

