% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/point.R
\name{point}
\alias{point}
\title{Create a point}
\usage{
point(x)
}
\arguments{
\item{x}{A pair of points, of the form \code{c(longitude,latitude)}}
}
\description{
Create a point
}
\examples{
point(c(40,-74.50))
}

