% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample.R
\name{random}
\alias{random}
\alias{sample}
\title{Return a FeatureCollection with given number of features at random}
\usage{
random(n = 10)

sample(features = NULL, n = 100)
}
\arguments{
\item{n}{(integer) Number of features to select.}

\item{features}{A FeatureCollection}
}
\description{
Return a FeatureCollection with given number of features at random
}
\examples{
# Random set of points
random(2)
random(10)
random(1000)

# Sample points from a FeatureCollection
dat <- points_average
sample(dat, 1)
sample(dat, 2)
sample(dat, 3)
}

