% Generated by roxygen2 (4.0.2): do not edit by hand
\name{random}
\alias{random}
\alias{sample}
\title{Return a FeatureCollection with given number of features at random}
\usage{
random(n = 10)

sample(features = NULL, n = 100)
}
\arguments{
\item{n}{(integer) Number of features to select.}

\item{features}{A FeatureCollection}
}
\description{
Return a FeatureCollection with given number of features at random
}
\examples{
# Random set of points
random(2)
random(10)
random(1000)

# Sample points from a FeatureCollection
sample(lawn_data$points_average, 1)
sample(lawn_data$points_average, 2)
sample(lawn_data$points_average, 3)
}

