% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\title{Inside}
\usage{
tag(points, polygons, polyId, containingPolyId)
}
\arguments{
\item{points}{Input point}

\item{polygons}{Input polygon or multipolygon}

\item{polyId}{property in polygons to add to joined Point features}

\item{containingPolyId}{property in points in which to store joined
property from polygons}
}
\value{
\code{TRUE} if the Point IS inside the Polygon, \code{FALSE} if
the Point IS NOT inside the Polygon
}
\description{
Takes a Point and a Polygon or MultiPolygon and determines if the point
resides inside the polygon. The polygon can be convex or concave. The
function accounts for holes
}
\examples{
\dontrun{
pts <- random(n = 30)
polys <- triangle_grid(c(-77.3876, 38.7198, -76.9482, 39.0277), 30, 'miles')
tag(pts, polys, 'fill', 'marker-color')
}
}

