% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\title{Visualize geojson}
\usage{
view(x)
}
\arguments{
\item{x}{Input, a geojson character string or list}
}
\value{
Opens a map with the geojson object(s)
}
\description{
Visualize geojson
}
\examples{
\dontrun{
# from character string
view(lawn_data$polygons_average)
view(lawn_data$filter_features)
view(lawn_data$polygons_within)
view(lawn_data$polygons_count)

# from a list
library("jsonlite")
mylist <- fromJSON(lawn_data$polygons_average, FALSE)
view(mylist)

# from json (a jsonlite class)
x <- minify(lawn_data$points_count)
class(x)
view(x)
}
}

