% Generated by roxygen2 (4.0.2): do not edit by hand
\name{within}
\alias{within}
\title{Return points that fall within polygons}
\usage{
within(points, polygons)
}
\arguments{
\item{points}{FeatureCollection of points}

\item{polygons}{FeatureCollection of polygons}
}
\description{
Return points that fall within polygons
}
\examples{
cat(lawn_data$points_within)
cat(lawn_data$polygons_within)
within(lawn_data$points_within, lawn_data$polygons_within)
}

